requirements.txt
python-telegram-bot[job-queue]==21.6
google-genai>=1.0.0,<2.0.0
gspread==6.1.2
google-auth==2.35.0
pandas==2.2.3
numpy==1.26.4
pyyaml==6.0.2
python-dotenv==1.0.1
aiofiles==23.2.1
aiohttp==3.10.5
aiosqlite==0.20.0
tenacity==9.0.0
pydantic-settings==2.5.2
pytz==2024.2
# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
# NOTE: apscheduler is NOT pinned here — python-telegram-bot[job-queue] manages its version
# DO NOT add: apscheduler or google-generativeai (deprecated EOL Nov 2025)
